Conflicts:
	res/xml/levels.xml
	src/com/example/ship/game/Ship.java

+++ src/com/example/ship/game/hud/GameHUD.java
@@ -266,7 +266,7 @@ public class GameHUD extends HUD {
 
     }
 
-    public void updateProgressBar(int progress) {
+    public void updateProgressBar(float progress) {
         progressBar.setProgress(progress);
     }
